created the score.js file and added webpack
